refactor merossclient namespace
